LayoutTests/fast/scrolling/scroll-animator-overlay-scrollbars-clicked-expected.txt

CONSOLE MESSAGE: line 18: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 18: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 20: MainFrameView: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 20: FrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 20: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 21: MainFrameView: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 22: MainFrameView: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 23: MainFrameView: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 23: FrameView: mouseExitedContentArea
CONSOLE MESSAGE: line 23: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=165056.
